+++ Lab1/check_connection.h
@@ -16,7 +16,7 @@ void setup_keepalive(int* sockfd);
 int check_connection(int* sockfd, FILE* logger);
 int write_with_check(int* sockfd, const char* buffer, int len, FILE* logger, FILE* f);
 int read_with_check(int* sockfd, char** buffer, int len, FILE* logger, FILE* f);
-int write_with_check_int(int* sockfd, long double* buffer, FILE* logger, FILE* f);
-int read_with_check_int(int* sockfd, long double* buffer, FILE* logger, FILE* f);
+int write_with_check_int(int* sockfd, long* buffer, FILE* logger, FILE* f);
+int read_with_check_int(int* sockfd, long* buffer, FILE* logger, FILE* f);
 
 #endif
